requirements.txt
python-rtmidi
jupyterlab
numpy
pyaudio
matplotlib
scipy
python-osc
sklearn
argcomplete